+++ scrapper/src/main/java/edu/java/services/jdbc/JdbcUserService.java
@@ -1,5 +1,7 @@
 package edu.java.services.jdbc;
 
+import edu.java.advice.NotFoundException;
+import edu.java.advice.UserAlreadyRegisteredException;
 import edu.java.domain.User;
 import edu.java.repositories.UserRepository;
 import edu.java.services.UserService;
@@ -17,12 +19,20 @@ public class JdbcUserService implements UserService {
 
     @Override
     public void add(long userId) {
+        User user = userRepository.getUser(userId);
+        if(user != null){
+            throw new UserAlreadyRegisteredException();
+        }
         userRepository.add(new User(userId));
 
     }
 
     @Override
     public void remove(long userId) {
+        User user = userRepository.getUser(userId);
+        if(user == null){
+            throw new NotFoundException();
+        }
         userRepository.remove(new User(userId));
     }
 }
